Add audits for barrel file risks, Rust panic risks, and secret candidates
- Implement tests for detecting large index barrel files, wildcard-heavy barrel files, and ignoring small index barrel files.
- Introduce Rust panic risk detection in AST with tests for handling unwraps in report renderers.
- Create finding construction and JWT token detection for secret candidates.
- Add rendering functions for context risk graphs in console, Markdown, DOT, and Mermaid formats.
- Implement visibility profile tests for findings and apply visibility profiles to scan summaries.
- Add support functions for risk overlays, including workspace hotspots and graph impacts.
