+++ CHANGELOG.md
@@ -19,7 +19,7 @@ The format is based on Keep a Changelog, and this project follows Semantic Versi
 
 ### Changed
 
-- Internal: split the 580-line `audits::code_quality::language_risk::pattern` god file by moving each language's AST node-emitter into the sibling pattern module that already owns that language's pattern data (`go`, `js`, `python`, `managed`), leaving only the line/AST dispatch and the shared finding-construction infrastructure (`RiskPattern`, `push_pattern_finding`) in `pattern.rs` (266 lines). Every source file in the module is now under 300 lines. Internal refactor only; no behavior, finding, or public API change.
+- Internal: split nine 300–580-line "god files" into focused submodules so every non-facade source file stays under 300 lines, with no behavior, finding, or public API change. `language_risk::pattern` moved each language's AST node-emitter into its sibling pattern module (`go`/`js`/`python`/`managed`), keeping only dispatch + the shared `RiskPattern`/`push_pattern_finding` infrastructure; `rust_panic_risk` extracted its AST candidate-collection into `ast`; `review::signals::behavioral::removed` extracted its AST walks into `removed_ast`; `commands::graph` moved its console/Markdown/DOT/Mermaid renderers into `graph::render`; `secret_candidate` split finding construction, masking, and JWT detection into an included `finding.rs`; `risk::overlays` moved its derivation helpers into `overlays::support`; and the compact console header detail lines into `compact::header`. Large inline test modules (`rust_panic_risk`, `architecture::barrel_file_risk`, `findings::visibility`) were also split into sibling test files. The four cohesive facade/algorithm files (`scan::scanner::file`, `scan::cache`, `graph`, `review`) are intentionally left as-is.
 - Internal: the `repopilot review` diff parser now captures each hunk's added and removed line text (`ChangedFile.hunks`), not just the changed-line ranges it derived before. This is the foundation for content-based review signals; `ranges` and all current output are unchanged and `hunks` is not serialized into the JSON report.
 - Internal: added a review-layer content bridge (`review::signals::content`) that re-reads a changed file's pre- and post-change source — from the working tree, or `git show <ref>:<path>` for a ref range — and pairs it with the shared tree-sitter parse view, so content-based signals can analyze each side of a change without keeping scan-time file content in memory.
 - Internal: `repopilot review` now unifies its boundary, behavioral, and algorithmic signals into three confidence tiers — *definitely sensitive*, *maybe sensitive*, and *large-diff-or-noise* — on `ReviewReport.tiered_signals`, wired through the review pipeline and reach-enriched from the coupling graph. A network call added inside an auth/request-trust file is escalated to *definitely*; a large diff with nothing flagged surfaces one volume note. `boundary_signals`/`boundary_missing_test` are unchanged; the tiered view is rendered across the console, Markdown, and JSON outputs (see the Added entry above).
