+++ src/risk/overlays.rs
@@ -1,13 +1,15 @@
 use crate::baseline::diff::{BaselineStatus, FindingBaselineStatus};
 use crate::baseline::key::normalized_relative_path;
-use crate::findings::types::{Finding, Severity};
-use crate::graph::{CouplingGraph, compute_metrics};
+use crate::findings::types::Finding;
+use crate::graph::CouplingGraph;
 use std::collections::{HashMap, HashSet};
-use std::path::{Component, Path, PathBuf};
+use std::path::{Path, PathBuf};
 
 use super::model::{
     FORMULA_VERSION, GraphImpact, RiskInputs, RiskSignal, clamp_score, priority_for_score, signal,
 };
+
+mod support;
 
 pub fn apply_baseline_overlay(
     findings: &mut [Finding],
@@ -55,7 +57,7 @@ pub fn apply_review_overlay(findings: &mut [Finding], in_diff: &[bool]) {
 }
 
 pub fn apply_workspace_hotspot_overlay(findings: &mut [Finding]) {
-    let hotspot_packages = workspace_hotspots(findings);
+    let hotspot_packages = support::workspace_hotspots(findings);
     if hotspot_packages.is_empty() {
         return;
     }
@@ -84,13 +86,13 @@ pub fn apply_workspace_hotspot_overlay(findings: &mut [Finding]) {
 }
 
 pub fn apply_graph_overlay(findings: &mut [Finding], graph: &CouplingGraph) {
-    let impacts = graph_impacts(graph);
+    let impacts = support::graph_impacts(graph);
     if impacts.is_empty() {
         return;
     }
 
     for finding in findings {
-        let Some(path) = finding_path_key(finding) else {
+        let Some(path) = support::finding_path_key(finding) else {
             continue;
         };
         let Some(impact) = impacts.get(path.as_str()).copied() else {
@@ -123,7 +125,7 @@ pub fn apply_blast_radius_overlay(
 ) {
     let impacted = blast_radius
         .iter()
-        .map(|path| path_key(path))
+        .map(|path| support::path_key(path))
         .collect::<HashSet<_>>();
     if impacted.is_empty() {
         return;
@@ -154,13 +156,13 @@ pub fn apply_blast_radius_overlay(
 }
 
 pub fn apply_cluster_overlay(findings: &mut [Finding]) {
-    let cluster_sizes = repeated_cluster_sizes(findings);
+    let cluster_sizes = support::repeated_cluster_sizes(findings);
     if cluster_sizes.is_empty() {
         return;
     }
 
     for finding in findings {
-        let Some(key) = finding_cluster_key(finding) else {
+        let Some(key) = support::finding_cluster_key(finding) else {
             continue;
         };
         let Some(size) = cluster_sizes.get(&key).copied().filter(|size| *size >= 3) else {
@@ -171,7 +173,7 @@ pub fn apply_cluster_overlay(findings: &mut [Finding]) {
             signal(
                 "cluster.repeated",
                 "cluster",
-                cluster_weight(size),
+                support::cluster_weight(size),
                 "same rule appears repeatedly in the same repository area",
             ),
             RiskInputs {
@@ -216,91 +218,5 @@ fn apply_overlay_signal(finding: &mut Finding, signal: RiskSignal, _inputs: Risk
     finding.risk.formula_version = FORMULA_VERSION.to_string();
 }
 
-fn workspace_hotspots(findings: &[Finding]) -> HashSet<String> {
-    let mut counts: HashMap<String, usize> = HashMap::new();
-    for finding in findings {
-        if finding.severity >= Severity::High
-            && let Some(package) = &finding.workspace_package
-        {
-            *counts.entry(package.clone()).or_default() += 1;
-        }
-    }
-    counts
-        .into_iter()
-        .filter_map(|(package, count)| (count > 1).then_some(package))
-        .collect()
-}
-
-fn graph_impacts(graph: &CouplingGraph) -> HashMap<String, GraphImpact> {
-    compute_metrics(graph)
-        .into_iter()
-        .filter_map(|metric| {
-            let impact = if metric.fan_in >= 3 || metric.fan_out >= 8 {
-                GraphImpact::Hub
-            } else if metric.fan_in >= 2 {
-                GraphImpact::Dependency
-            } else {
-                return None;
-            };
-            Some((path_key(&metric.path), impact))
-        })
-        .collect()
-}
-
-fn repeated_cluster_sizes(findings: &[Finding]) -> HashMap<String, usize> {
-    let mut counts = HashMap::new();
-    for finding in findings {
-        if let Some(key) = finding_cluster_key(finding) {
-            *counts.entry(key).or_default() += 1;
-        }
-    }
-    counts.retain(|_, count| *count >= 3);
-    counts
-}
-
-fn finding_cluster_key(finding: &Finding) -> Option<String> {
-    let path = finding
-        .evidence
-        .first()
-        .map(|evidence| path_key(&evidence.path))?;
-    Some(format!("{}:{}", finding.rule_id, cluster_scope(&path)))
-}
-
-fn finding_path_key(finding: &Finding) -> Option<String> {
-    finding
-        .evidence
-        .first()
-        .map(|evidence| path_key(&evidence.path))
-}
-
-fn cluster_scope(path: &str) -> String {
-    let mut parts = path.split('/').filter(|part| !part.is_empty());
-    match (parts.next(), parts.next()) {
-        (Some(first), Some(second)) if second.contains('.') => first.to_string(),
-        (Some(first), Some(second)) => format!("{first}/{second}"),
-        (Some(first), None) => first.to_string(),
-        _ => ".".to_string(),
-    }
-}
-
-fn path_key(path: &Path) -> String {
-    path.components()
-        .filter_map(|component| match component {
-            Component::CurDir => None,
-            Component::Normal(value) => Some(value.to_string_lossy().to_string()),
-            Component::RootDir | Component::Prefix(_) | Component::ParentDir => {
-                Some(component.as_os_str().to_string_lossy().to_string())
-            }
-        })
-        .collect::<Vec<_>>()
-        .join("/")
-}
-
-fn cluster_weight(size: usize) -> i16 {
-    match size {
-        0..=2 => 0,
-        3..=5 => 3,
-        6..=15 => 5,
-        _ => 7,
-    }
-}
+// Derivation helpers (workspace hotspots, graph impacts, cluster sizing, path
+// keys) live in the `support` submodule.
